changhe to float

+++ python/samples/stockD.py
@@ -27,11 +27,11 @@ def get_value(identifier):
     value = subprocess.Popen(['curl', '-s', get_value_url], stdout=subprocess.PIPE).communicate()[0]
     j = json.loads(value[5:len(value)-2])
     returnString = [str(j['t']),str(j['l']),str(j['c']),str(j['cp'])]
-    if int(j['c']) == 0:
+    if float(j['c']) == 0.00:
         returnString.append("blue")
-    if int(j['c']) > 0:
+    if float(j['c']) > 0.00:
         returnString.append("green")
-    if int(j['c']) < 0:
+    if float(j['c']) < 0.00:
         returnString.append("red")    
     return returnString
